requirements.txt
flask
blinker
peewee
pymysql
python-decouple
redis
celery
cerberus
requests
ujson
pyDes
redis-py-cluster

ipython
Flask-Script
gunicorn
gevent
raven
supervisor
flasgger
Flask-FlatPages
flower
